Wait to center map until we get user's location when app is active
If the app wakes up from a significant location change and the map is
centered immediately, then the user will see the old location when they
enter the app.

Even if we delay centering the map until the user opens the app, they
will still see the old location because the map centering code will run
before the new location is fetched.

Therefore, we wait until the app is active before sending out the
active location and triggering the map centering.